requirements.txt
numpy
matplotlib
seaborn
Pillow
tqdm
prettytable
opencv-python
torch
PyYAML
dill
wandb
plotly
multiprocess